Add enrichment to sync and don't default to create zip file

+++ resources/contrib/sync/sync_to_usb.ps1
@@ -2,16 +2,18 @@
 $SOURCE_DIRECTORY="$HOME\Documents\workspace\dfirws"
 $CURRENT="$PWD"
 
-Set-Location "$SOURCE_DIRECTORY"
-
-Start-Process -Wait 'C:\Program Files\7-Zip\7z.exe' -Argumentlist "a dfirws.zip .\downloads .\mount .\setup"
-Get-Job | Receive-Job
-Get-Job | Remove-Job
-Copy-Item dfirws.zip $CURRENT
-Remove-Item -Force dfirws.zip
-Set-Location "$CURRENT"
+if ($args -contains "--zip") {
+    Set-Location "$SOURCE_DIRECTORY"
+    Start-Process -Wait 'C:\Program Files\7-Zip\7z.exe' -Argumentlist "a dfirws.zip .\downloads .\enrichment .\mount .\setup"
+    Get-Job | Receive-Job
+    Get-Job | Remove-Job
+    Copy-Item dfirws.zip $CURRENT
+    Remove-Item -Force dfirws.zip
+    Set-Location "$CURRENT"
+}
 
 Robocopy.exe $SOURCE_DIRECTORY\downloads .\dfirws\downloads /MIR /MT:96
+Robocopy.exe $SOURCE_DIRECTORY\enrichment .\dfirws\downloads /MIR /MT:96
 Robocopy.exe $SOURCE_DIRECTORY\mount .\dfirws\mount /MIR /MT:96
 Robocopy.exe $SOURCE_DIRECTORY\setup .\dfirws\setup /MIR /MT:96
